Refactor ML processor modals
