+++ public/pages/workflow_detail/workflow_inputs/processor_inputs/ml_processor_inputs/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './ml_processor_inputs';
+export * from './modals';
