+++ public/pages/workflow_detail/workflow_inputs/processor_inputs/ml_processor_inputs/modals/configure_prompt_modal.tsx
@@ -38,11 +38,11 @@ import {
   PromptPreset,
   WorkflowFormValues,
   customStringify,
-} from '../../../../../../common';
+} from '../../../../../../../common';
 import {
   parseModelInputs,
   parseModelInputsObj,
-} from '../../../../../utils/utils';
+} from '../../../../../../utils/utils';
 
 interface ConfigurePromptModalProps {
   config: IProcessorConfig;
