+++ public/pages/workflow_detail/workflow_inputs/processor_inputs/ml_processor_inputs/modals/index.ts
@@ -6,3 +6,4 @@
 export * from './input_transform_modal';
 export * from './output_transform_modal';
 export * from './configure_prompt_modal';
+export * from './override_query_modal';
