+++ public/pages/workflow_detail/workflow_inputs/processor_inputs/ml_processor_inputs/modals/input_transform_modal.tsx
@@ -48,7 +48,7 @@ import {
   customStringify,
   REQUEST_PREFIX,
   REQUEST_PREFIX_WITH_JSONPATH_ROOT_SELECTOR,
-} from '../../../../../../common';
+} from '../../../../../../../common';
 import {
   formikToPartialPipeline,
   generateTransform,
@@ -56,20 +56,20 @@ import {
   getInitialValue,
   prepareDocsForSimulate,
   unwrapTransformedDocs,
-} from '../../../../../utils';
+} from '../../../../../../utils';
 import {
   searchIndex,
   simulatePipeline,
   useAppDispatch,
-} from '../../../../../store';
-import { getCore } from '../../../../../services';
+} from '../../../../../../store';
+import { getCore } from '../../../../../../services';
 import {
   generateArrayTransform,
   getDataSourceId,
   parseModelInputs,
   parseModelInputsObj,
-} from '../../../../../utils/utils';
-import { BooleanField, MapArrayField } from '../../input_fields';
+} from '../../../../../../utils/utils';
+import { BooleanField, MapArrayField } from '../../../input_fields';
 
 interface InputTransformModalProps {
   uiConfig: WorkflowConfig;
